feat: add abstract bridge class
