+++ src/bridges/usdcBridge.ts
@@ -1,233 +1,78 @@
 import {
-  BigNumberish,
-  BytesLike,
-  ethers,
   AbiCoder,
   ContractTransactionResponse,
   Overrides,
+  Contract,
 } from 'ethers';
-import {
-  Address,
-  PaymasterParams,
-  PriorityOpResponse,
-  TransactionResponse,
-} from '../types';
+import {Address, FinalizeL1DepositParams, TransactionLike} from '../types';
 import {IL1Nullifier__factory} from '../typechain';
-import {
-  REQUIRED_L1_TO_L2_GAS_PER_PUBDATA_LIMIT,
-  checkBaseCost,
-  scaleGasLimit,
-} from '../utils';
 import {Wallet} from '../wallet';
-import {insertGasPrice} from '../adapters';
+import {IDepositTransaction, IWithdrawTransaction, Bridge} from './bridge';
 
-interface IDepositTransaction {
-  token: Address;
-  amount: BigNumberish;
-  to?: Address;
-  operatorTip?: BigNumberish;
-  bridgeAddress: Address;
-  approveERC20?: boolean;
-  l2GasLimit?: BigNumberish;
-  gasPerPubdataByte?: BigNumberish;
-  refundRecipient?: Address;
-  overrides?: Overrides;
-  approveOverrides?: Overrides;
-}
+const L1_BRIDGE_ABI = ['function L1_USDC_TOKEN() view returns (address)'];
 
-interface IWithdrawTransaction {
-  token: Address;
-  amount: BigNumberish;
-  to?: Address;
-  bridgeAddress: Address;
-  paymasterParams?: PaymasterParams;
-  overrides?: Overrides;
-  approveERC20?: boolean;
-  approveOverrides?: Overrides;
-}
+const L2_BRIDGE_ABI = ['function L2_USDC_TOKEN() view returns (address)'];
 
-export class USDCBridge {
-  constructor(protected readonly wallet: Wallet) {}
+/**
+ * `USDCBridge` is an implementation of the Bridge abstract class which provides methods
+ * for depositing and withdrawing USDC tokens using Custom USDC Bridge that can be found here:
+ * https://github.com/matter-labs/usdc-bridge
+ */
+export class USDCBridge extends Bridge {
+  constructor(wallet: Wallet) {
+    super(wallet);
+  }
 
-  private getSecondBridgeDepositCalldata(
+  override async validateDepositParams(
     transaction: IDepositTransaction
-  ): string {
+  ): Promise<void> {
+    const l1Token = new Contract(
+      transaction.bridgeAddress,
+      L1_BRIDGE_ABI,
+      this.wallet._signerL1()
+    );
+    const bridgeL1Token = await l1Token.L1_USDC_TOKEN();
+
+    if (bridgeL1Token.toLowerCase() !== transaction.token.toLowerCase()) {
+      throw new Error('Wrong token address for USDC bridge deposit.');
+    }
+  }
+
+  override async getSecondBridgeDepositCalldata(
+    transaction: IDepositTransaction
+  ): Promise<string> {
     return AbiCoder.defaultAbiCoder().encode(
       ['address', 'uint256', 'address'],
       [transaction.token, transaction.amount, transaction.to]
     );
   }
-  /**
-   * @inheritDoc
-   *
-   * @example Deposit USDC.
-   *
-   * import { Wallet, Provider, types, USDCBridge } from "zksync-ethers";
-   * import { ethers } from "ethers";
-   *
-   * const PRIVATE_KEY = "<WALLET_PRIVATE_KEY>";
-   * const USDC_TOKEN_L1_ADDRESS = "<USDC_TOKEN_ADDRESS>";
-   * const USDC_BRIDGE_L1_ADDRESS = "<USDC_BRIDGE_L1_ADDRESS>";
-   * const AMOUNT = "5";
-   *
-   * const provider = Provider.getDefaultProvider(types.Network.Sepolia);
-   * const ethProvider = ethers.getDefaultProvider("sepolia");
-   * const wallet = new Wallet(PRIVATE_KEY, provider, ethProvider);
-   * const usdcBridge = new USDCBridge(wallet);
-   *
-   * const depositTx = await usdcBridge.deposit({
-   *   token: USDC_TOKEN_L1_ADDRESS,
-   *   amount: ethers.parseUnits(AMOUNT, 6),
-   *   approveERC20: true,
-   *   bridgeAddress: USDC_BRIDGE_L1_ADDRESS,
-   * });
-   *
-   * // Note that we wait not only for the L1 transaction to complete but also for it to be
-   * // processed by zkSync. If we want to wait only for the transaction to be processed on L1,
-   * // we can use `await depositTx.waitL1Commit()`
-   * await depositTx.wait();
-   */
-  async deposit(transaction: IDepositTransaction): Promise<PriorityOpResponse> {
-    const chainId = (await this.wallet._providerL2().getNetwork()).chainId;
-    const bridgehub = await this.wallet.getBridgehubContract();
 
-    const tx = {
-      ...transaction,
-      to: transaction.to || (await this.wallet.getAddress()),
-      operatorTip: transaction.operatorTip || 0,
-      gasPerPubdataByte:
-        transaction.gasPerPubdataByte ||
-        REQUIRED_L1_TO_L2_GAS_PER_PUBDATA_LIMIT,
-      overrides: transaction.overrides || {},
-    };
-    const secondBridgeCalldata = await this.getSecondBridgeDepositCalldata(tx);
-    if (!tx.l2GasLimit) {
-      tx.l2GasLimit = await this.wallet._getL2GasLimitFromCustomBridge({
-        ...tx,
-        customBridgeData: secondBridgeCalldata,
-      });
-    }
-    await insertGasPrice(this.wallet._providerL1(), tx.overrides);
-
-    const gasPriceForEstimation =
-      tx.overrides.maxFeePerGas || tx.overrides.gasPrice;
-    const baseCost = await bridgehub.l2TransactionBaseCost(
-      chainId,
-      gasPriceForEstimation as BigNumberish,
-      tx.l2GasLimit,
-      tx.gasPerPubdataByte
+  override async validateWithdrawParams(
+    transaction: IWithdrawTransaction
+  ): Promise<void> {
+    const l2Token = new Contract(
+      transaction.bridgeAddress,
+      L2_BRIDGE_ABI,
+      this.wallet._signerL2()
     );
+    const bridgeL2Token = await l2Token.L2_USDC_TOKEN();
 
-    const mintValue = baseCost + BigInt(tx.operatorTip);
-    tx.overrides.value = tx.overrides.value || mintValue;
-    tx.overrides.from = await this.wallet.getAddress();
-
-    await checkBaseCost(baseCost, mintValue);
-
-    const depositTx =
-      await bridgehub.requestL2TransactionTwoBridges.populateTransaction(
-        {
-          chainId,
-          mintValue,
-          l2Value: 0,
-          l2GasLimit: tx.l2GasLimit,
-          l2GasPerPubdataByteLimit: tx.gasPerPubdataByte,
-          refundRecipient: tx.refundRecipient ?? ethers.ZeroAddress,
-          secondBridgeAddress: tx.bridgeAddress,
-          secondBridgeValue: 0,
-          secondBridgeCalldata,
-        },
-        tx.overrides
-      );
-    if (tx.approveERC20) {
-      const allowance = await this.wallet.getAllowanceL1(
-        tx.token,
-        tx.bridgeAddress
-      );
-      if (allowance < BigInt(transaction.amount)) {
-        const approveTx = await this.wallet.approveERC20(tx.token, tx.amount, {
-          bridgeAddress: tx.bridgeAddress,
-          ...tx.approveOverrides,
-        });
-        await approveTx.wait();
-      }
+    if (bridgeL2Token.toLowerCase() !== transaction.token.toLowerCase()) {
+      throw new Error('Wrong token address for USDC bridge withdrawal.');
     }
-
-    if (!depositTx.gasLimit) {
-      const baseGasLimit = await this.wallet
-        ._providerL1()
-        .estimateGas(depositTx);
-      depositTx.gasLimit = scaleGasLimit(baseGasLimit);
-    }
-
-    return await this.wallet.getPriorityOpResponse(
-      await this.wallet._signerL1().sendTransaction(depositTx)
-    );
   }
 
-  /**
-   * @inheritDoc
-   *
-   * @example Withdraw USDC.
-   *
-   * import { Wallet, Provider, types, USDCBridge } from "zksync-ethers";
-   *
-   * const PRIVATE_KEY = "<WALLET_PRIVATE_KEY>";
-   * const USDC_TOKEN_L2_ADDRESS = "<USDC_TOKEN_L2_ADDRESS>";
-   * const USDC_BRIDGE_L2_ADDRESS = "<USDC_BRIDGE_L2_ADDRESS>";
-   * const AMOUNT = "5";
-   *
-   * const provider = Provider.getDefaultProvider(types.Network.Sepolia);
-   * const wallet = new Wallet(PRIVATE_KEY, provider);
-   * const usdcBridge = new USDCBridge(wallet);
-   *
-   * const withdrawTx = await usdcBridge.withdraw({
-   *   token: USDC_TOKEN_L2_ADDRESS,
-   *   amount: ethers.parseUnits(AMOUNT, 6),
-   *   bridgeAddress: USDC_BRIDGE_L2_ADDRESS,
-   *   approveERC20: true,
-   * });
-   */
-  async withdraw(
-    transaction: IWithdrawTransaction
-  ): Promise<TransactionResponse> {
-    if (transaction.approveERC20) {
-      const allowance = await this.wallet.getAllowanceL2(
-        transaction.token,
-        transaction.bridgeAddress
-      );
-      if (allowance < BigInt(transaction.amount)) {
-        const approveTx = await this.wallet.approveERC20L2(
-          transaction.token,
-          transaction.amount,
-          {
-            ...transaction.approveOverrides,
-            bridgeAddress: transaction.bridgeAddress,
-          }
-        );
-        await approveTx.wait();
-      }
-    }
-
-    const walletAddress = await this.wallet.getAddress();
-    const tx: IWithdrawTransaction & {
-      from: Address;
-      to: Address;
-      overrides: Overrides;
-    } = {
-      ...transaction,
-      from: walletAddress,
-      to: transaction.to || walletAddress,
-      overrides: transaction.overrides || {},
-    };
+  override async populateWithdrawTransaction(
+    tx: IWithdrawTransaction
+  ): Promise<TransactionLike> {
     const bridge = await this.wallet
       ._providerL2()
       .connectL2Bridge(tx.bridgeAddress!);
     const populatedWithdrawTx = await bridge.withdraw.populateTransaction(
-      tx.to,
-      tx.token,
+      tx.to!,
+      tx.token!,
       tx.amount,
-      tx.overrides
+      tx.overrides!
     );
 
     if (tx.paymasterParams) {
@@ -236,40 +81,19 @@ export class USDCBridge {
       };
     }
 
-    return this.wallet.sendTransaction(populatedWithdrawTx);
+    return populatedWithdrawTx;
   }
 
-  /**
-   * @inheritDoc
-   *
-   * @example
-   *
-   * import { Wallet, Provider, types, utils } from "zksync-ethers";
-   * import { ethers } from "ethers";
-   *
-   * const PRIVATE_KEY = "<WALLET_PRIVATE_KEY>";
-   *
-   * const provider = Provider.getDefaultProvider(types.Network.Sepolia);
-   * const ethProvider = ethers.getDefaultProvider("sepolia");
-   * const wallet = new Wallet(PRIVATE_KEY, provider, ethProvider);
-   *
-   * const WITHDRAWAL_HASH = "<WITHDRAWAL_TX_HASH>";
-   * const finalizeWithdrawTx = await wallet.finalizeWithdrawal(WITHDRAWAL_HASH);
-   */
-  async finalizeWithdrawal(
+  protected override async finalizeL1Deposit(
     bridgeAddress: Address,
-    withdrawalHash: BytesLike,
-    index = 0,
-    overrides: Overrides = {}
+    finalizeParams: FinalizeL1DepositParams,
+    overrides?: Overrides
   ): Promise<ContractTransactionResponse> {
     const bridgeContract = IL1Nullifier__factory.connect(
       bridgeAddress,
       this.wallet._signerL1()
     );
-    const finalizeParams = await this.wallet.getFinalizeDepositParams(
-      withdrawalHash,
-      index
-    );
+
     return await bridgeContract.finalizeWithdrawal(
       finalizeParams.chainId,
       finalizeParams.l2BatchNumber,
